bigfile fix

+++ www/wiki/js/app/controller/bigfileController.js
@@ -10,31 +10,42 @@ define([
     app.registerController("bigfileController", ["$scope", function ($scope) {
         var qiniuBack;
         const biteToG = 1024*1024*1024;
+        const ErrFilenamePatt = new RegExp('^[^\\\\/\*\?\|\<\>\:\"]+$');
         $scope.selectedType = "图片";
         if((util.getPathname() !="/wiki/user_center")){
             $scope.isModal=true;
         }
-        $scope.cancel = function () {
+        $scope.cancel = function (params) {
             if ($scope.uploadingFiles && $scope.uploadingFiles.length > 0 && !$scope.finishUploading){
                 console.log("正在上传");
-                config.services.confirmDialog({
-                    "title": "关闭提示",
+                var confirmObj = {
+                    "title": "提示",
                     "confirmBtnClass": "btn-danger",
                     "theme": "danger",
                     "content": "还有文件正在上传，确定关闭窗口？"
-                }, function () {
-                    $scope.$dismiss();
+                };
+                if (params){
+                    confirmObj.content = "还有文件正在上传，请完成后重试，或者打开新窗口操作";
+                    confirmObj.cancelBtn = false;
+                    confirmObj.confirmBtnClass = "";
+                }
+                config.services.confirmDialog(confirmObj, function () {
+                    if (params){
+                        return;
+                    }
+                    $scope.$dismiss(params);
                 });
             }else{
-                $scope.$dismiss();
+                $scope.$dismiss(params);
             }
 
         };
 
         var getFileByUsername = function () {
-            util.post(config.apiUrlPrefix + "bigfile/getByUsername",{}, function(data){
+            util.post(config.apiUrlPrefix + "bigfile/getByUsername",{pageSize:100000}, function(data){
                 data = data || {};
                 $scope.filelist = data.filelist;
+                $scope.filesCount = data.total;
             });
         };
 
@@ -54,7 +65,7 @@ define([
         };
 
         var isExceed = function (unUsed, uploadingSize) {
-            if (unUsed > uploadingSize){
+            if (!unUsed || !uploadingSize || unUsed > uploadingSize){
                 return false;
             } else {
                 config.services.confirmDialog({
@@ -68,7 +79,10 @@ define([
             }
         };
 
-        $scope.initQiniu = function(){
+        $scope.initQiniu = function(type){
+            if (type !== "isUpdating"){
+                $scope.updatingFile = {};
+            }
             if ($scope.finishUploading){
                 $scope.uploadingFiles = [];
             }else if ($scope.finishUploading === false){
@@ -107,7 +121,7 @@ define([
                         	    data.map(function (file) {
                                     conflictFileName.push(file.filename);
                                     contentHtml+='<p class="dialog-info"><span class="text-success glyphicon glyphicon-ok"></span> '+ file.filename +'</p>';
-                                    conflictSize += file.size;
+                                    conflictSize += file.size || 0;
                                 });
 
                         	    if (isExceed($scope.storeInfo.unUsed * biteToG, (filesSize - conflictSize))){
@@ -204,6 +218,14 @@ define([
                         });
                     },
                     'Error': function(up, err, errTip) {
+                        if (err.code == -601){
+                            option.filters = {};
+                            $scope.uploadingFiles = $scope.uploadingFiles || [];
+                            $scope.uploadingFiles.push(err.file);
+                            up.files.push(err.file);
+                            up.start();
+                            return;
+                        }
                         //上传出错时，处理相关的事情
                         $scope.uploadingFiles = [];
                         $scope.updatingFile = [];
@@ -253,7 +275,9 @@ define([
         $scope.$watch("$viewContentLoaded", init);
 
         $scope.stopUpload = function (file) {
-            qiniuBack.stop();
+            if (file.status == 2){
+                qiniuBack.stop();
+            }
             config.services.confirmDialog({
                 "title": "取消上传",
                 "confirmBtnClass": "btn-danger",
@@ -261,10 +285,15 @@ define([
                 "content": "确定取消该文件上传吗？"
             }, function () {
                 $scope.storeInfo.unUsed += file.size/biteToG;
-                qiniuBack.removeFile(file);
                 file.isDelete = true;
+                qiniuBack.removeFile(file.id);
+                if (qiniuBack.files.length <=0){
+                    $scope.uploadingFiles = [];
+                }
+                qiniuBack.start();
+            }, function () {
+                qiniuBack.start();
             });
-            qiniuBack.start();
         };
 
         $scope.deleteFile = function(files, index) {
@@ -274,25 +303,32 @@ define([
                 "theme":"danger",
                 "content":"确定删除文件吗？"
             },function(){
-                if (index && !Array.isArray(files)){
+                if (!Array.isArray(files)){
                     var file = files;
                     file.index = index;
                     files = [];
                     files.push(file);
                 }
                 var fnList = [];
+                var deleteFileSize = 0;
                 for (var i=files.length -1; i >= 0;i--){
                     fnList.push((function (index) {
                         return function (finish) {
                             util.post(config.apiUrlPrefix + 'bigfile/deleteById', {_id:files[index]._id}, function (data) {
-                                $scope.filelist[files[index].index].isDelete = true;
+                                files[index].isDelete = true;
+                                $scope.filesCount--;
+                                deleteFileSize += files[index].file.size;
+                                finish();
                             }, function (err) {
                                 console.log(err);
+                                finish();
                             });
                         }
                     })(i));
-                    util.batchRun(fnList);
                 }
+                util.batchRun(fnList, function () {
+                    $scope.storeInfo.unUsed += deleteFileSize/biteToG;
+                });
             });
         };
 
@@ -300,13 +336,31 @@ define([
             var deletingArr = $scope.filelist.filter(function (file) {
                 return file.index >= 0;
             });
-            $scope.deleteFile(deletingArr);
+            if (deletingArr.length <= 0){
+                config.services.confirmDialog({
+                    "title": "提示",
+                    "content": "请至少选择一个要删除的文件！",
+                    "cancelBtn": false
+                }, function () {
+                });
+            }else{
+                $scope.deleteFile(deletingArr);
+            }
         };
 
         var changeFileName = function (file, filename, targetElem) {
-            targetElem.attr("contenteditable", "false");
             $scope.nameErr="";
-            // filename = filename.trim();
+
+            if (!filename || file == ""){
+                targetElem.html(file.filename);
+                config.services.confirmDialog({
+                    "title": "重命名失败",
+                    "content": "文件名不能为空！",
+                    "cancelBtn": false
+                }, function () {
+                });
+                return;
+            }
 
             var newFileSplit = filename.split(".");
             var newExt = newFileSplit[newFileSplit.length-1];
@@ -329,22 +383,32 @@ define([
                 newExt = oldExt;
             }
 
-            util.post(config.apiUrlPrefix+"bigfile/changeFilename",{
-                "_id": file._id,
-                "filename": filename
-            }, function (result) {
-                console.log(result);
-            }, function (err) {
-                console.log(err);
+            util.post(config.apiUrlPrefix + "bigfile/getByFilenameList", {filelist:[filename]}, function(data){
+                if (data.length > 0){
+                    targetElem.html(file.filename);
+                    config.services.confirmDialog({
+                        "title": "重命名失败",
+                        "content": "网盘中已存在该文件名！",
+                        "cancelBtn": false
+                    }, function () {
+                    });
+                }else{
+                    util.post(config.apiUrlPrefix+"bigfile/changeFilename",{
+                        "_id": file._id,
+                        "filename": filename
+                    }, function (result) {
+                        file.filename = filename;
+                    }, function (err) {
+                        console.log(err);
+                    });
+                }
             });
         };
 
         $scope.updateFile = function (file) {
-            // console.log("更新功能开发中。。。");
             $scope.updatingFile = file;
             $("#activeUpload").tab("show");
-            $scope.initQiniu();
-            // Message.info("更新功能开发中。。。");
+            $scope.initQiniu("isUpdating");
         };
         
         var removeAllTags = function (str) {
@@ -361,13 +425,22 @@ define([
             targetElem.focus();
             targetElem.bind("blur", function () {
                 var filename = targetElem.html();
-                filename = filename.replace(/[\n|<br>]/g, "").trim();
-                targetElem.html(filename);
+                filename = filename.replace(/\n|<br>|&nbsp;/g, "").trim();
+                if (!ErrFilenamePatt.test(filename)){
+                    targetElem.html(file.filename);
+                    config.services.confirmDialog({
+                        "title": "重命名失败",
+                        "content": '文件名不能包含下列任何字符：\\\\ / : * ? " < > |',
+                        "cancelBtn": false
+                    }, function () {
+                    });
+                    return;
+                }
                 changeFileName(file, filename, targetElem);
+                targetElem.attr("contenteditable", "false");
             });
             targetElem.on("paste", function () { // contenteditable中粘贴会包含html结构
                 setTimeout(function () {
-                    console.log(targetElem.html());
                     targetElem.html(removeAllTags(targetElem.html()));
                 });
             })
@@ -378,7 +451,6 @@ define([
                 _id:file._id,
             }, function(data){
                 if (data) {
-                    console.log(data);
                     var a = document.createElement('a');
                     var url = data;
                     a.href = url;
@@ -401,7 +473,16 @@ define([
                 });
             }
 
-            $scope.$dismiss(files);
+            if (files.length <= 0){
+                config.services.confirmDialog({
+                    "title": "提示",
+                    "content": "请至少选择一个要插入的文件！",
+                    "cancelBtn": false
+                }, function () {
+                });
+            }else{
+                $scope.cancel(files);
+            }
         };
 
         $scope.insertFilesUrl = function () {
@@ -427,7 +508,7 @@ define([
                 default:
                     break;
             }
-            $scope.$dismiss({
+            $scope.cancel({
                 "type": type,
                 "url": url
             });
